+++ main/src/main/java/net/oneandone/stool/util/ServerXml.java
@@ -55,7 +55,7 @@ public class ServerXml {
         file.writeXml(document);
     }
 
-    public void configure(Ports ports, String url, KeyStore keystore, boolean cookies, Stage stage) throws XmlException {
+    public void configure(Ports ports, String url, KeyStore keystore, boolean cookies, Stage stage, boolean http2) throws XmlException {
         Element template;
         Element service;
 
@@ -66,7 +66,7 @@ public class ServerXml {
                 service = (Element) template.cloneNode(true);
                 document.getDocumentElement().appendChild(service);
                 service(service, vhost);
-                connectors(service, vhost, keystore);
+                connectors(service, vhost, keystore, http2);
                 contexts(stage, vhost.context(hostname, url), vhost.httpPort(), service, cookies, vhost.docroot.join("WEB-INF"));
             }
         }
@@ -102,15 +102,15 @@ public class ServerXml {
         host.insertBefore(element, host.getFirstChild());
     }
 
-    private void connectors(Element service, Vhost host, KeyStore keyStore) {
+    private void connectors(Element service, Vhost host, KeyStore keyStore, boolean http2) {
         String ip;
 
         ip = "0.0.0.0";
         try {
             connectorDisable(service, "Connector[starts-with(@protocol,'AJP')]");
-            connectorEnable(service, HTTP_PATH, ip, host.httpPort(), host.httpsPort());
+            connectorEnable(service, HTTP_PATH, ip, host.httpPort(), host.httpsPort(), http2);
             if (keyStore != null) {
-                sslConnector(service, HTTPS_PATH, host.httpsPort(), ip, keyStore);
+                sslConnector(service, HTTPS_PATH, host.httpsPort(), ip, keyStore, http2);
             } else {
                 connectorDisable(service, HTTPS_PATH);
             }
@@ -119,7 +119,7 @@ public class ServerXml {
         }
     }
 
-    private void connectorEnable(Element service, String path, String ip, int port, int sslport) throws XmlException {
+    private void connectorEnable(Element service, String path, String ip, int port, int sslport, boolean http2) throws XmlException {
         Element element;
 
         element = selector.element(service, path);
@@ -127,6 +127,17 @@ public class ServerXml {
         element.setAttribute("address", ip);
         element.setAttribute("useBodyEncodingForURI", "true");
         element.setAttribute("redirectPort", Integer.toString(sslport));
+        http2(element, http2);
+    }
+
+    private void http2(Element element, boolean enable) {
+        if (enable) {
+            Element http2;// TODO: only Tomcat 8.5+
+            // https://tomcat.apache.org/tomcat-8.5-doc/config/http2.html
+            http2 = element.getOwnerDocument().createElement("UpgradeProtocol");
+            http2.setAttribute("className", "org.apache.coyote.http2.Http2Protocol");
+            element.appendChild(http2);
+        }
     }
 
     private void connectorDisable(Element service, String path) throws XmlException {
@@ -139,7 +150,7 @@ public class ServerXml {
     }
 
 
-    private void sslConnector(Element service, String path, int port, String ip, KeyStore keystore) throws XmlException {
+    private void sslConnector(Element service, String path, int port, String ip, KeyStore keystore, boolean http2) throws XmlException {
         Element element;
 
         element = selector.elementOpt(service, path);
@@ -164,6 +175,7 @@ public class ServerXml {
 
         element.removeAttribute("SSLCertificateFile");
         element.removeAttribute("SSLCertificateKeyFile");
+        http2(element, http2);
 
     }